Cleanup and organize a little the code
