+++ src/global.ts
@@ -1,0 +1,3 @@
+import * as os from "node:os";
+
+export const hostname = os.hostname();
